requirements-extra.txt
# Trading APIs & Exchanges
python-binance==1.0.19
ccxt>=4.1.0
websockets>=12.0

# AI Integration
openai>=1.3.0
transformers>=4.35.0

# Web Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0
flask-socketio>=5.3.0

# Database & Caching
redis>=5.0.1

# Environment Configuration
python-dotenv>=1.0.0

# Logging
loguru>=0.7.2

# System Monitoring
gpustat>=1.0.0

# Testing
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Technical Analysis
ta-lib>=0.4.28
tulipy>=0.4.0

# Portfolio Optimization
riskfolio-lib>=4.1.0
empyrical>=0.5.0

# Sentiment Analysis
textblob>=0.17.1
vaderSentiment>=3.3.2

# Blockchain & On-chain
web3>=6.0.0

# Alternative ML Models
xgboost>=2.0.0
catboost>=1.2.0
